Baseline: Add basic working dataloader

+++ src/dataloader.h
@@ -2,12 +2,12 @@
 #ifndef DATALOADER_H
 #define DATALOADER_H
 
-// Load data from a binary file or a CSV file
+// Load NxD data from a binary file or a CSV file
 // Returns 0 on success, -1 on failure
-int load_data_bin(const char* filename, float** data, int* n, int* d);
-int load_data_csv(const char* filename, float** data, int* n, int* d);
+int load_data_bin(const char* filename, double** data, int* n, int* d);
+int load_data_csv(const char* filename, double** data, int* n, int* d);
 
 // Free the allocated data
-void free_data(float* data);
+void free_data(double* data);
 
 #endif // DATALOADER_H
